Use Uint8Array for writeTests rgb test
With the new writeArrayBuffer defaulting to Float64 writing, the
"writeTests > should write rgb data with sensible defaults" was failing
on the StripByteCount, since it no longer equals 27 with a regular
Array. I had switched it to the new value, 216, but decided it's likely
better to not change the expect() statements when introducing changes.